:recycle: refactor: make primary keys bigint

+++ server/DataAccess/Interfaces/IRefreshTokenRepository.cs
@@ -2,6 +2,6 @@ namespace server.DataAccess.Interfaces;
 
 public interface IRefreshTokenRepository
 {
-    Task<bool> IsRequestTokenValid(int userId, string refreshToken);
-    Task StoreToken(int userId, DateTime refreshExpiry, string refreshToken);
+    Task<bool> IsRequestTokenValid(long userId, string refreshToken);
+    Task StoreToken(long userId, DateTime refreshExpiry, string refreshToken);
 }
